Fix a signed/unsigned warning introduced in 1b4568c
vout counter must be unsigned.

(cherry picked from commit 5bb765507b394e1fe205355b0b9937d95f8a39e6)

# Conflicts:
#	src/wallet.cpp

+++ src/wallet.cpp
@@ -838,7 +838,7 @@ void CWalletTx::GetAmounts(list<COutputEntry>& listReceived,
     }
 
     // Sent/received.
-    for (int i = 0; i < vout.size(); ++i)
+    for (unsigned int i = 0; i < vout.size(); ++i)
     {
         const CTxOut& txout = vout[i];
 
@@ -869,7 +869,8 @@ void CWalletTx::GetAmounts(list<COutputEntry>& listReceived,
             address = CNoDestination();
         }
 
-        COutputEntry output = {address, txout.nValue, i};
+        COutputEntry output = {address, txout.nValue, (int)i};
+
         // If we are debited by the transaction, add the output as a "sent" entry
         if (nDebit > 0)
             listSent.push_back(output);
